+++ example/pl_grid_complex_example.dart
@@ -1,50 +1,15 @@
-# pl_grid
-
-The PlGrid is a grid view for showing data in a table with a sticky header and pagination components.
-
-## Getting Started
-
-To add the grid on your project, install it by adding the pl_grid package dependency to your pubspec.yaml file, then import the PlGrid class like so: 
-```dart
+import 'package:flutter/material.dart';
 import 'package:pl_grid/pl_grid.dart';
-````
-Add the widget like in the example down below:
 
-```dart
-PlGrid(
-    headerColumns: ['Id', 'Name', 'Age'],
-    data: [
-        [1, 'Bruno', 34],
-        [2, 'Lindsey', 39],
-        [3, 'Roberto', 18],
-        [4, 'Yasmin', 22]
-    ],
-    columnWidthsPercentages: [15, 70, 15],
-    curPage: 1,
-    maxPages: 4,
-    width: 300,
-    height: 200,
-)
-```
+class PlGridExample extends StatefulWidget {
+  @override
+  _PlGridExampleState createState() => _PlGridExampleState();
+}
 
-and it will produce the following:
-
-<img src="https://github.com/playlinesdev/pl_grid/blob/master/sample1.png?raw=true"/>
-
-One can also use other parameters for example paginationItemClick that is a function that receives the index of the page touched/clicked to execute any routine such as call api with next page results
-
-The property asCard is set as default to true, but could be set to false, producing the following:
-
-<img src="https://github.com/playlinesdev/pl_grid/blob/master/sample1.png?raw=true"/>
-
-<img src="https://github.com/playlinesdev/pl_grid/blob/master/sample2.png?raw=true"/>
-<img src="https://github.com/playlinesdev/pl_grid/blob/master/sample_web.png?raw=true"/>
-
-
-A complex example to show the many features that can be achieved with this package
-
-```dart
-Scaffold(
+class _PlGridExampleState extends State<PlGridExample> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
       backgroundColor: Colors.black38,
       body: Padding(
         padding: const EdgeInsets.only(top: 80, left: 80),
@@ -118,7 +83,5 @@ Scaffold(
         ),
       ),
     );
-```
-
-<img src="https://github.com/playlinesdev/pl_grid/blob/master/sample_complex.png?raw=true"/>
-<img src="https://github.com/playlinesdev/pl_grid/blob/master/sample_complex1.png?raw=true"/>
+  }
+}